Refactor signUp.tsx to sign-up.tsx
